+++ src/registrardb-internal.cc
@@ -159,11 +159,11 @@ void RegistrarDbInternal::doClear(const MsgSip& msg, const shared_ptr<ContactUpd
 	}
 
 	LOGI << "AOR " << key << " found";
-	shared_ptr<Record> r = (*it).second;
+	const auto record = (*it).second;
 
 	mRecords.erase(it);
 	mLocalRegExpire.remove(key);
-	listener->onRecordFound(nullptr);
+	listener->onRecordFound(record);
 }
 
 void RegistrarDbInternal::clearAll() {